refactor: content descriptions add
