+++ core/network/build.gradle.kts
@@ -1,7 +1,7 @@
 plugins {
     id("android-core-convention")
     alias(libs.plugins.compose.compiler)
-    id("kotlinx-serialization") // TODO: libs
+    alias(libs.plugins.kotlin.serialization)
 }
 
 android {
