+++ core/theme/src/main/java/ru/gribbirg/theme/material/MaterialColorsSchemes.kt
@@ -1,7 +1,29 @@
-package ru.gribbirg.theme
+package ru.gribbirg.theme.material
 
 import androidx.compose.material3.darkColorScheme
 import androidx.compose.material3.lightColorScheme
+import ru.gribbirg.theme.Black
+import ru.gribbirg.theme.DarkBlue
+import ru.gribbirg.theme.DarkGray
+import ru.gribbirg.theme.DarkGrayLight
+import ru.gribbirg.theme.DarkGreen
+import ru.gribbirg.theme.DarkOverlay
+import ru.gribbirg.theme.DarkPrimaryBack
+import ru.gribbirg.theme.DarkPrimaryLabel
+import ru.gribbirg.theme.DarkRed
+import ru.gribbirg.theme.DarkSecondaryBack
+import ru.gribbirg.theme.DarkTertiaryLabel
+import ru.gribbirg.theme.LightBlue
+import ru.gribbirg.theme.LightGray
+import ru.gribbirg.theme.LightGrayLight
+import ru.gribbirg.theme.LightGreen
+import ru.gribbirg.theme.LightOverlay
+import ru.gribbirg.theme.LightPrimaryBack
+import ru.gribbirg.theme.LightPrimaryLabel
+import ru.gribbirg.theme.LightRed
+import ru.gribbirg.theme.LightSecondaryBack
+import ru.gribbirg.theme.LightTertiaryLabel
+import ru.gribbirg.theme.White
 
 internal val lightColorSchemeImpl = lightColorScheme(
     primary = LightBlue,
