+++ core/theme/src/main/java/ru/gribbirg/theme/material/MaterialTypography.kt
@@ -1,10 +1,11 @@
-package ru.gribbirg.theme
+package ru.gribbirg.theme.material
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.sp
+import ru.gribbirg.theme.custom.robotoFontFamily
 
 internal val typography = Typography(
     displayLarge = TextStyle(
