+++ core/ui/src/main/java/ru/gribbirg/ui/previews/Templates.kt
@@ -8,8 +8,8 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import ru.gribbirg.theme.AppTheme
-import ru.gribbirg.theme.TodoAppTheme
+import ru.gribbirg.theme.custom.AppTheme
+import ru.gribbirg.theme.custom.TodoAppTheme
 
 /**
  * Templates for previews
