+++ feature/edit/src/main/java/ru/gribbirg/edit/components/EditScreenAppBarComponent.kt
@@ -33,7 +33,7 @@ import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
 import kotlinx.coroutines.launch
-import ru.gribbirg.theme.AppTheme
+import ru.gribbirg.theme.custom.AppTheme
 import ru.gribbirg.todoapp.edit.R
 import ru.gribbirg.ui.components.CloseButton
 import ru.gribbirg.ui.previews.BooleanPreviewParameterProvider
