+++ feature/list/src/main/java/ru/gribbirg/list/components/TodoItemListCollapsingToolbarComponent.kt
@@ -37,7 +37,7 @@ import me.onebone.toolbar.CollapsingToolbarScaffold
 import me.onebone.toolbar.ScrollStrategy
 import me.onebone.toolbar.rememberCollapsingToolbarScaffoldState
 import ru.gribbirg.list.TodoItemsListUiState
-import ru.gribbirg.theme.AppTheme
+import ru.gribbirg.theme.custom.AppTheme
 import ru.gribbirg.todoapp.list.R
 import ru.gribbirg.ui.previews.DefaultPreview
 import ru.gribbirg.ui.previews.LanguagePreviews
@@ -165,7 +165,7 @@ internal fun TodoItemListCollapsingToolbar(
                 IconButton(onClick = toSettingsScreen) {
                     Icon(
                         Icons.Default.Settings,
-                        contentDescription = null, // TODO
+                        contentDescription = stringResource(id = R.string.settings),
                         tint = AppTheme.colors.blue,
                     )
                 }
