+++ feature/list/src/main/java/ru/gribbirg/list/components/TodoItemRowSwipeBackgroundComponent.kt
@@ -16,7 +16,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import ru.gribbirg.theme.AppTheme
+import ru.gribbirg.theme.custom.AppTheme
 
 /**
  * Background for swipe to dismiss
