+++ feature/settings/src/main/res/values-b+ru/strings.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="login">Войти</string>
+    <string name="logout">Выйти</string>
 </resources>
